# HW 11/ClinicService/ClinicDesctop.Data/OpenAPIs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "ClinicService",
      "version": "1.0"
    },
    "paths": {
      "/api/Client/create": {
        "post": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientCreate",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClientRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClientRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateClientRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Client/update": {
        "put": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientUpdate",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateClientRequest"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateClientRequest"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateClientRequest"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Client/delete": {
        "delete": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientDelete",
          "parameters": [
            {
              "name": "clientId",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "integer",
                    "format": "int32"
                  }
                }
              }
            }
          }
        }
      },
      "/api/Client/get-all": {
        "get": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientGetAll",
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Client"
                    }
                  }
                },
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Client"
                    }
                  }
                },
                "text/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Client"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/Client/get/{clientId}": {
        "get": {
          "tags": [
            "Client"
          ],
          "operationId": "ClientGetById",
          "parameters": [
            {
              "name": "clientId",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success",
              "content": {
                "text/plain": {
                  "schema": {
                    "$ref": "#/components/schemas/Client"
                  }
                },
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Client"
                  }
                },
                "text/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Client"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Client": {
          "type": "object",
          "properties": {
            "clientId": {
              "type": "integer",
              "format": "int32"
            },
            "document": {
              "type": "string",
              "nullable": true
            },
            "surName": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "patronymic": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "CreateClientRequest": {
          "type": "object",
          "properties": {
            "document": {
              "type": "string",
              "nullable": true
            },
            "surName": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "patronymic": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
        "UpdateClientRequest": {
          "type": "object",
          "properties": {
            "clientId": {
              "type": "integer",
              "format": "int32"
            },
            "document": {
              "type": "string",
              "nullable": true
            },
            "surName": {
              "type": "string",
              "nullable": true
            },
            "firstName": {
              "type": "string",
              "nullable": true
            },
            "patronymic": {
              "type": "string",
              "nullable": true
            },
            "birthday": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        }
      }
    }
  }